Split up console display implementations in different files.
