+++ README.md
@@ -48,12 +48,12 @@ Here is a minimal example that constructs a 100 by 100 display colored red:
 #![feature(generic_const_exprs)]
 
 use console_display::{
-    console_display::PixelDisplay,
     display_driver::DisplayDriver,
     pixel::color_pixel::{
         self,
         ColorOctPixel,
     },
+    pixel_display::StaticPixelDisplay,
 };
 
 fn main() {
@@ -61,10 +61,9 @@ fn main() {
     // (8 pixels per console character, 4 high, 2 wide).
     // Set dimensions to 100 by 100 pixels (50 by 25 characters)
     // with a red fill.
-    let disp: PixelDisplay<ColorOctPixel> =
-        PixelDisplay::<ColorOctPixel>::new::<100, 100>(
-            color_pixel::RGBColor { r: 255, g: 0, b: 0 },
-        );
+    let disp = StaticPixelDisplay::<ColorOctPixel, 100, 100>::new(
+        color_pixel::RGBColor { r: 255, g: 0, b: 0 },
+    );
 
     // Wrap the display in a driver to manage interactions with the terminal
     // like resizing, enabling raw mode, providing an update loop.
