+++ src/display.rs
@@ -1,2 +1,4 @@
+pub mod character_display;
 pub mod console_display;
 pub mod display_driver;
+pub mod pixel_display;
